fix: close the last high-severity findings, and every remaining one is medium
IR-013 (high, the last one) — completion validated the staged key and promoted
afterwards. Because a presigned PUT stays valid until it expires, a client could
replace the object in between, so the promoted bytes could differ from the
metadata that was checked and every later view, download and packet would serve
the replacement. Promotion now happens first and the object is validated where it
now lives; a rejection after the move deletes the promoted copy instead of
orphaning it, and a racing completion is answered 409.

IR-004 (high) — the privileges migration created a role named `rgm_app`, as
though it were the application's own, and granted it only `events`. A deployment
that believed the name could not read `users`, `sessions` or `active_memberships`
and could not start. It is now `rgm_runtime`, holding exactly what the
application does, with `events` still append-only by grant as well as by trigger.
Migration 005 migrates an existing installation and is deliberately tolerant:
roles are cluster-scoped and grants are not, so a role another database still
depends on is left alone rather than failing every deployment that has one.
Verified by assuming the role on a real PostgreSQL 17.

IR-008 (high) — `rgm login` wrote a long-lived token with the process umask and
accepted it in argv, where `ps` shows it. Now 0600 inside a 0700 directory, with
the token read from RGM_TOKEN or --token-stdin.

IR-009 (high) — project selection is global and the output path comes from the
bug number, so `rgm pull 1` in one repository could extract another project's
BUG-1 over the same directory. `pull` now checks the packet's own meta.json and
refuses a packet from a different project before writing anything.

IR-017 (med) — /api/invites/redeem was not CSRF-exempt, so a signed-in user was
refused while an anonymous one succeeded: exactly backwards.

Also: byte_size was a bigint, which node-postgres returns as a *string*, so the
API sent "byteSize": "64" against PostgreSQL and 64 against the embedded driver.
Uploads cap at 8 MB, so the column is an integer now (migration 006) — the same
trap as count(*), but from a schema choice rather than a query.

All 13 high-severity findings are closed. 15 medium remain, all recorded in
IR-FINDINGS.md.

Suite: 287 passing on the embedded driver, through the pg driver, and against
real PostgreSQL 17. Smoke run green.

+++ db/migrations/001_init.sql
@@ -181,7 +181,7 @@ CREATE TABLE bug_attachments (
   storage_key  text NOT NULL UNIQUE,
   -- Tester-supplied. Kept as DATA only; never used as a path on disk (RGM3-007).
   filename     text NOT NULL,
-  byte_size    bigint NOT NULL CHECK (byte_size > 0),
+  byte_size    integer NOT NULL CHECK (byte_size > 0),
   content_type text NOT NULL CHECK (content_type LIKE 'image/%'),
   uploaded_at  timestamptz NOT NULL DEFAULT now(),
   FOREIGN KEY (project_id, bug_id)
